Debug and some modularization

+++ uninstall.sh
@@ -2,9 +2,6 @@
 
 set -eu
 
-export STEAM_HOME='/home/steam'
-export ECHO_DONE='echo -e \e[32mdone\e[0m.'
-
 read -p "This will uninstall Killing Floor 2 on this machine. Press 'y' to continue: " -n 1 -r
 echo
 if [[ ! $REPLY =~ ^[Yy]$ ]]
@@ -13,30 +10,27 @@ then
     exit
 fi
 
+export ROOT="${BASH_SOURCE%/*}"
+
+source "${ROOT}/lib/globals.sh"
+source "${ROOT}/lib/aliases.sh"
+source "${ROOT}/lib/flags.sh"
+source "${ROOT}/lib/functions.sh"
+
 echo 'Uninstalling Killing Floor 2... '
 
 # Essentially, this file should be bin/* undone, in reverse order.
-
-function safe_erase ()
-{
-    RET=1
-    yum -q list installed $1 >/dev/null 2>&1 && RET=$?
-    if [ ${RET} -eq 0 ]
-    then
-        yum -y -q erase $1
-    fi
-}
 
 echo -n 'Stopping KF2 services... '
 
 if [ -f /etc/systemd/system/kf2.service ]
 then
-    systemctl --quiet stop kf2.service
+    systemctl ${SYSTEMCTL_FLAGS} stop kf2.service
 fi
 
 if [ -f /etc/systemd/system/kf2autokick.service ]
 then
-    systemctl --quiet stop kf2autokick.service
+    systemctl ${SYSTEMCTL_FLAGS} stop kf2autokick.service
 fi
 
 ${ECHO_DONE}
@@ -49,8 +43,7 @@ ${ECHO_DONE}
 
 # Autokick
 echo -n 'Removing auto-kick bot... '
-safe_erase nodejs
-safe_erase yarn
+kf2_yum_erase nodejs yarn
 rm -f /etc/yum.repos.d/nodesource-el7.repo
 rm -f /etc/yum.repos.d/yarn.repo
 rm -rf ${STEAM_HOME}/kf2autokick
@@ -58,14 +51,13 @@ ${ECHO_DONE}
 
 # Config generator
 echo -n 'Removing config generator... '
-safe_erase crudini
-safe_erase epel-release
+kf2_yum_erase crudini epel-release
 ${ECHO_DONE}
 
 # Backup
 DATE_STR=$(date +%Y%m%d-%H%M%S)
 BACKUP_FILE="${STEAM_HOME}/Config-${DATE_STR}.tgz"
-echo -n "Backing up current KF2 config as ${BACKUP_FILE}... "
+echo -ne "Backing up current KF2 config as \e[36m${BACKUP_FILE}\e[0m... "
 rm -f ${STEAM_HOME}/Config/Internal
 sudo -u steam sh -c "tar czfh ${BACKUP_FILE} -C ${STEAM_HOME} Config"
 ${ECHO_DONE}
@@ -75,10 +67,10 @@ echo -n 'Removing firewall rules... '
 
 if [ -f /etc/firewalld/services/kf2.xml ]
 then
-    firewall-cmd --quiet --remove-service=kf2 --permanent
-    firewall-cmd --quiet --reload
-    firewall-cmd --quiet --delete-service=kf2 --permanent
-    firewall-cmd --quiet --reload
+    firewall-cmd ${FIREWALLCMD_FLAGS} --remove-service=kf2 --permanent
+    firewall-cmd ${FIREWALLCMD_FLAGS} --reload
+    firewall-cmd ${FIREWALLCMD_FLAGS} --delete-service=kf2 --permanent
+    firewall-cmd ${FIREWALLCMD_FLAGS} --reload
 fi
 
 ${ECHO_DONE}
@@ -92,41 +84,39 @@ ${ECHO_DONE}
 # Systemd
 echo -n 'Removing KF2 services... '
 
-systemctl --quiet daemon-reload
+systemctl ${SYSTEMCTL_FLAGS} daemon-reload
 
 if [ -f /etc/systemd/system/kf2.service ]
 then
-    systemctl --quiet disable kf2.service
+    systemctl ${SYSTEMCTL_FLAGS} disable kf2.service
 fi
 
 if [ -f /etc/systemd/system/kf2autokick.service ]
 then
-    systemctl --quiet disable kf2autokick.service
+    systemctl ${SYSTEMCTL_FLAGS} disable kf2autokick.service
 fi
 
 rm -f /etc/systemd/system/kf2.service
 rm -f /etc/systemd/system/kf2autokick.service
 rm -rf /etc/systemd/system/kf2.service.d
-systemctl --quiet daemon-reload
+systemctl ${SYSTEMCTL_FLAGS} daemon-reload
 
 ${ECHO_DONE}
 
 # Steam + KF2
 echo -n 'Removing KF2 and Steam... '
-rm -rf ${STEAM_HOME}/Steam
-rm -rf ${STEAM_HOME}/.steam
+if [ ${SKIP_KFGAME} -ne 1 ]
+then
+    rm -rf ${STEAM_HOME}/Steam
+    rm -rf ${STEAM_HOME}/.steam
+fi
 rm -f ${STEAM_HOME}/Cache
 rm -f ${STEAM_HOME}/Workshop
 ${ECHO_DONE}
 
 # Deps
 echo -n 'Removing dependencies... '
-
-for PKG in glibc.i686 libstdc++.i686 unzip dos2unix patch
-do
-    safe_erase $PKG
-done
-
+kf2_yum_erase glibc.i686 libstdc++.i686 unzip dos2unix patch
 ${ECHO_DONE}
 
 # the only thing we don't remove is the steam user, because it has user config
